將Puzzle.cs分成多個檔案
