+++ ExclusiveProgram/Control.cs
@@ -1,29 +1,21 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.ComponentModel;
-using System.Data;
 using System.Drawing;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using System.Windows.Forms;
 using Emgu.CV;
-using Emgu.CV.CvEnum;
 using Emgu.CV.Structure;
-using Puzzle.Concrete;
-using Puzzle.Framework;
-using RASDK.Arm;
-using RASDK.Basic;
-using RASDK.Basic.Message;
-using RASDK.Gripper;
+using ExclusiveProgram.puzzle.visual.concrete;
+using ExclusiveProgram.puzzle.visual.framework;
+using ExclusiveProgram.ui.component;
 
 namespace ExclusiveProgram
 {
     public partial class Control : MainForm.ExclusiveControl
     {
-        private readonly PuzzleRecognizer recognizer;
-        private readonly PuzzleLocator locator;
-        private readonly PuzzleCorrector corrector;
+        private readonly IPuzzleRecognizer recognizer;
+        private readonly IPuzzleLocator locator;
+        private readonly IPuzzleCorrector corrector;
 
         public Control()
         {
@@ -37,12 +29,6 @@ namespace ExclusiveProgram
 
         private void button1_Click(object sender, EventArgs e)
         {
-            AA();
-        }
-
-        private void AA()
-        {
-
             var factory = new DefaultPuzzleFactory((int)numericUpDown_threshold.Value,locator,recognizer,corrector);
             var image = VisualSystem.LoadImageFromFile("samples\\Test.jpg");
 
@@ -53,14 +39,15 @@ namespace ExclusiveProgram
                 //Point StartPoint = new Point((int)(location.Coordinate.X - (location.Size.Width / 2.0f)), (int)(location.Coordinate.Y - (location.Size.Height / 2.0f)));
                 //CvInvoke.Rectangle(showImage, new Rectangle(StartPoint.X, StartPoint.Y, location.Size.Width, location.Size.Height), new MCvScalar(0, 0, 255), 3);
                 //CvInvoke.PutText(showImage, "Angle:" + location.Angle, new Point(StartPoint.X, StartPoint.Y - 15),FontFace.HersheySimplex,2, new MCvScalar(0, 0, 255),3);
-                
-                var p = new PictureBox();
-                p.Image = VisualSystem.Mat2Image<Bgr>(result.image).ToBitmap();
-                p.Size= new Size(200,200);
-                p.SizeMode=PictureBoxSizeMode.StretchImage;
-                puzzleView.Controls.Add(p);
+                var co= new UserControl1();
+                co.setImage(VisualSystem.Mat2Image<Bgr>(result.image).ToBitmap());
+                co.setLabel(result.Angel, result.coordinate);
+                //p.Image = VisualSystem.Mat2Image<Bgr>(result.image).ToBitmap();
+                puzzleView.Controls.Add(co);
+
                 Console.WriteLine("" + result.position);
             }
         }
+
     }
 }
