+++ ExclusiveProgram/puzzle.visual/framework/IPuzzleCorrector.cs
@@ -1,0 +1,15 @@
+﻿using Emgu.CV;
+using Emgu.CV.Structure;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace ExclusiveProgram.puzzle.visual.framework
+{
+    public interface IPuzzleCorrector
+    {
+        Image<Bgr,byte> Correct(Image<Bgr, byte> input,LocationResult raw);
+    }
+}
