+++ ExclusiveProgram/puzzle.visual/framework/IPuzzleFactory.cs
@@ -1,0 +1,16 @@
+﻿using Emgu.CV;
+using Emgu.CV.Structure;
+using ExclusiveProgram.puzzle.visual.framework;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace ExclusiveProgram.puzzle.visual.concrete
+{
+    public interface IPuzzleFactory
+    {
+        List<Puzzle_sturct> Execute(Image<Bgr, byte> input);
+    }
+}
